Consider m_hcycletotal an uint()

+++ src/ClassSimZ80.cpp
@@ -67,8 +67,8 @@ bool ClassSimZ80::initChip()
 // XXX Remove this timer here and implement it somewhere else (?)
 void ClassSimZ80::onTimeout()
 {
-    qreal hz = (m_hcyclecnt / 2.0) / (m_elapsed.elapsed() / 1000.0);
-    qDebug() << "Half-Cycles:" << m_hcycletotal << "~" << ((hz < 100) ? 0 : hz) << " Hz";
+    qreal hz = (uint(m_hcyclecnt) / 2.0) / (m_elapsed.elapsed() / 1000.0);
+    qDebug() << "Half-Cycles:" << uint(m_hcycletotal) << "~" << ((hz < 100) ? 0 : hz) << " Hz";
     if (m_runcount <= 0)
         m_timer.stop();
 }
